Moved hudson/plugins/report/jck/main/* to .../main/diff
